move fixture from test file to fixture fike
