+++ .pre-commit-config.yaml
@@ -27,7 +27,7 @@ repos:
     hooks:
       - id: flake8
         args: ["--max-line-length=120"]
-        exclude: tests/conftest.py # ignore pytest fixtures
+        exclude: up42/__init__.py|tests/conftest.py # ignore pytest fixtures
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: "v1.1.1" # Use the sha / tag you want to point at
     hooks:
