+++ tests/conftest.py
@@ -25,7 +25,7 @@ from .fixtures.fixtures_jobcollection import (
     jobcollection_single_mock,
 )
 from .fixtures.fixtures_jobtask import jobtask_live, jobtask_mock
-from .fixtures.fixtures_order import order_live, order_mock
+from .fixtures.fixtures_order import order_live, order_mock, order_parameters
 from .fixtures.fixtures_project import project_live, project_mock, project_mock_max_concurrent_jobs
 from .fixtures.fixtures_storage import storage_live, storage_mock
 from .fixtures.fixtures_tasking import (
