+++ tests/test_order.py
@@ -125,31 +125,6 @@ def test_get_assets_live(auth_live, order_parameters):
     assert assets[0].asset_id == ASSET_ORDER_ID
 
 
-@pytest.fixture
-def order_parameters():
-    return {
-        "dataProduct": "4f1b2f62-98df-4c74-81f4-5dce45deee99",
-        "params": {
-            "id": "aa1b5abf-8864-4092-9b65-35f8d0d413bb",
-            "aoi": {
-                "type": "Polygon",
-                "coordinates": [
-                    [
-                        [13.357031, 52.52361],
-                        [13.350981, 52.524362],
-                        [13.351544, 52.526326],
-                        [13.355284, 52.526765],
-                        [13.356944, 52.525067],
-                        [13.357257, 52.524409],
-                        [13.357031, 52.52361],
-                    ]
-                ],
-            },
-        },
-        "tags": ["Test", "SDK"],
-    }
-
-
 def test_place_order(order_parameters, auth_mock, order_mock, requests_mock):
     requests_mock.post(
         url=f"{API_HOST}/workspaces/{auth_mock.workspace_id}/orders",
